+++ crates/vancouver/src/check.rs
@@ -617,7 +617,7 @@ struct Receipt {
 
 #[allow(clippy::too_many_lines)]
 pub fn do_check(args: &crate::CheckArgs) -> Result<ExitCode, Error> {
-    let dependencies = crate::metadata::get_dependencies(&args.lock)?;
+    let dependencies = crate::metadata::get_dependencies(args.manifest.as_deref())?;
     if dependencies.is_empty() {
         return Err(Error::EmptyDependencies);
     }